requirements_build.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
face-recognition>=1.3.0
firebase-admin>=6.2.0
pyinstaller>=5.13.0
matplotlib>=3.7.0
accelerate>=0.20.0
tensorflow>=2.13.0
pyasn1>=0.4.8
protobuf>=4.21.0
cachetools>=5.3.0
attrs>=23.1.0
pydantic>=2.0.0

